missing int16_t (#14)

+++ KAV_Simulation/KAV_A3XX_EFIS_LCD.cpp
@@ -144,7 +144,7 @@ void KAV_A3XX_EFIS_LCD::displayDigit(uint8_t address, uint8_t digit)
     refreshLCD(address);
 }
 
-void KAV_A3XX_EFIS_LCD::set(int8_t messageID, char *setPoint)
+void KAV_A3XX_EFIS_LCD::set(int16_t messageID, char *setPoint)
 {
     int32_t data = atoi(setPoint);
     /* **********************************************************************************
